Revert "schedulers: add changefilter on nightly builds"
This reverts commit 705f02050e494769b94a4b2cea370d1016e52d96.

This is not correct and nightly builds have not triggered on repository
changes since.

Note that the following change still does not fix it:
@@ -46,8 +46,7 @@ def scheduler_nightly(name, builders, template_dfl,
codebases, hour, minute):
         cb = codebases.get(c_cb)
         return cb != None \
             and cb.get('branch') == c.properties.getProperty('branch')
\
-            and (cb.get('revision') == None \
-                or cb.get('revision') ==
		 c.properties.getProperty('revision'))
+            and cb.get('revision') == ""

     return schedulers.Nightly(
         name=name,
@@ -58,7 +57,7 @@ def scheduler_nightly(name, builders, template_dfl,
codebases, hour, minute):
         builderNames=builders,
         hour=hour,
         minute=minute,
-        change_filter=filter_nightly,
+        change_filter=util.ChangeFilter(filter_fn=filter_nightly),
         onlyIfChanged=True)

filter_fn is not easy to grasp, even having the filter function return
True does not trigger builds when changes are reported by the gitpoller.

This was to avoid redundant builds, but the load is pretty light for
now, so build more often than necessary instead of not.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>