schedulers: restore changefilter on nightly builds
Fix the confusion on the argument to the filter_fn.

Compare each Change object against the scheduler's given codebases (also
used as 'codebases' argument to fill the blanks). Automated build should
occur only if the change relates to a codebase and branch tracked by
that scheduler that is not on a preset revision.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>